pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "streamlang"
version = "1.1.0"
description = "A statically typed toy language compiled to Python."
authors = [{name = "Mohamed Ahmed", email = "hazalkoom048@gmail.com"}]
requires-python = ">=3.10"
dependencies = [
    "antlr4-python3-runtime==4.13.1",
    "pytest"
]

[tool.setuptools.packages.find]
where = ["src"]